backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
httpx==0.28.1
telethon==1.38.1
PyGithub==2.5.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
